requirements.txt
asgiref==3.5.2
boto3==1.24.67
botocore==1.27.67
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.1
Naked==0.1.31
pycryptodome==3.15.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
